api/requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.30.0
sqlalchemy[asyncio]>=2.0
asyncpg>=0.30.0
pydantic>=2.0
pydantic-settings>=2.0
boto3>=1.35.0
python-multipart>=0.0.9
pyyaml>=6.0
semver>=3.0
python-frontmatter>=1.1
httpx>=0.27.0
pytest>=8.0
pytest-asyncio>=0.24.0
aiosqlite>=0.20.0
moto[s3]>=5.0.0
